django/requirements.txt
asgiref==3.2.3
cffi==1.14.0
Django==3.0.3
django-cors-headers==3.2.1
django-webpack-loader==0.7.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
Pillow>=8.0
pycparser==2.20
PyJWT==1.7.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
